answered data questions
practiced while Sean was giving lecture. was able to produce code

for the most part while he pseudocoded.

once I knew how to implement it in code, with steps, I knew what to do.

Right now I feel the getting from a to b in steps is hard.

It also would have been good if I understood how to read python testing

or at the very least how to correct an error when one was returned.